Change Thread::instruction_counter type to uint64_t (from unsigned)

+++ include/viua/thread.h
@@ -70,7 +70,7 @@ class Thread {
     std::string return_message;     // message set by exception
 
     // FIXME: change unsigned to uint64_t
-    unsigned instruction_counter;
+    uint64_t instruction_counter;
     byte* instruction_pointer;
 
     std::queue<Type*> message_queue;
@@ -253,7 +253,7 @@ class Thread {
         }
 
         byte* begin();
-        inline unsigned counter() { return instruction_counter; }
+        inline uint64_t counter() { return instruction_counter; }
 
         inline std::tuple<int, std::string, std::string> exitcondition() {
             return std::tuple<int, std::string, std::string>(return_code, return_exception, return_message);
